+++ test/unit/animation.js
@@ -288,4 +288,32 @@
 
   });
 
+  $.each({
+    'positive':           { start:  3, reach: 15, expect: { speed:  1 }},
+    'negative':           { start: 13, reach:  5, expect: { speed: -1 }},
+    'positive over edge': { start: 20, reach:  3, expect: { speed:  1 }},
+    'negative over edge': { start:  2, reach: 18, expect: { speed: -1 }}
+  },
+  function(name, def){
+    asyncTest( 'Reach: Trigger `reach` event to playback to given frame - '+name, function()
+    {
+      expect(3);
+      var
+        $reel= $('#image').reel({
+          frame: def.start,
+          frames: 24,
+          speed: 1
+        })
+
+      $reel.trigger('reach', def.reach);
+
+      $(document).bind('stop.test', function(){
+        ok( true, 'Reached a stop');
+        equal( $reel.reel('frame'), def.reach, 'On the right frame');
+        equal( $reel.reel('speed'), def.expect.speed, 'With the right speed/direction');
+        start();
+      });
+    });
+  });
+
 })(jQuery);